requirements.txt
streamlit==1.38.0
pandas==2.2.3
python-dotenv==1.0.1
openai==1.52.0
numpy==2.1.2
